refactor : 별점, header경로 리팩토링
