+++ weco-app/components/Header/index.tsx
@@ -1,23 +1,19 @@
 import Image from 'next/image';
 import Router from 'next/router';
+import { HeaderRouter } from './HeaderRouter';
 
 export function MainHeader() {
-  const goWrite = () => {
-    Router.push('write');
-  };
-  const goLogin = () => {
-    Router.push('login');
-  };
   return (
     <nav className="m-auto max-w-7xl flex justify-between items-center h-20 py-0 px-2.5">
-      <Image src="/img/weco.png" width="150" height="150" />
+      <Image
+        src="/img/weco.png"
+        width="150"
+        height="150"
+        onClick={() => Router.push('/')}
+      />
       <div className="flex gap-8 items-center">
-        <button className="text-xl font-semibold" onClick={goWrite}>
-          새글 쓰기
-        </button>
-        <button className="text-xl font-semibold" onClick={goLogin}>
-          로그인
-        </button>
+        <HeaderRouter route="write" text="새글쓰기" />
+        <HeaderRouter route="login" text="로그인" />
       </div>
     </nav>
   );
